vector_service/requirements.txt
fastapi>=0.99.1,<1.0
uvicorn[standard]>=0.21.1
psycopg2-binary>=2.9.6
pydantic>=1.10.10,<2
